services/vlm/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pillow==10.4.0
torch==2.3.1
transformers==4.44.2
einops==0.8.0
accelerate==0.33.0
